Add files via upload
Generating random data & small dataframe similar to characteristics off Fortune500 dataset.

My goal for this notebook is to use the GROUPBY method on this small dataset followed by AGGREGATION methods to investigate the affect of such methods on the data. This approach makes for easier interpretation of these important methods as they can be difficult to interpret when first working with and especially when working with large datasets attempting to gain insight and understanding of the data.
The following notebook handles a small number of features and numeric that can be easily analyzed to improve understanding of these available methods in Python when working with pandas dataframes!